all current tests pass and style fixes
