+++ src/main/java/org/opensearch/ad/model/ValidationAspect.java
@@ -67,4 +67,3 @@ public enum ValidationAspect implements Name {
         return Name.getNameFromCollection(names, ValidationAspect::getName);
     }
 }
-
