+++ src/main/java/org/opensearch/ad/rest/RestValidateAnomalyDetectorAction.java
@@ -24,33 +24,31 @@
  * permissions and limitations under the License.
  */
 
-
 package org.opensearch.ad.rest;
 
-import com.google.common.collect.ImmutableList;
+import static org.opensearch.ad.util.RestHandlerUtils.TYPE;
+import static org.opensearch.ad.util.RestHandlerUtils.VALIDATE;
+import static org.opensearch.common.xcontent.XContentParserUtils.ensureExpectedToken;
+
+import java.io.IOException;
+import java.util.List;
+import java.util.Locale;
+
 import org.opensearch.ad.AnomalyDetectorPlugin;
 import org.opensearch.ad.constant.CommonErrorMessages;
 import org.opensearch.ad.model.AnomalyDetector;
 import org.opensearch.ad.settings.EnabledSetting;
 import org.opensearch.ad.transport.ValidateAnomalyDetectorAction;
 import org.opensearch.ad.transport.ValidateAnomalyDetectorRequest;
-
-import static org.opensearch.ad.util.RestHandlerUtils.*;
-import static org.opensearch.ad.util.RestHandlerUtils.DETECTOR_ID;
-import static org.opensearch.common.xcontent.XContentParserUtils.ensureExpectedToken;
-
 import org.opensearch.client.node.NodeClient;
 import org.opensearch.cluster.service.ClusterService;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.common.xcontent.XContentParser;
 import org.opensearch.rest.BaseRestHandler;
-import org.opensearch.rest.RestHandler;
 import org.opensearch.rest.RestRequest;
 import org.opensearch.rest.action.RestToXContentListener;
 
-import java.io.IOException;
-import java.util.List;
-import java.util.Locale;
+import com.google.common.collect.ImmutableList;
 
 public class RestValidateAnomalyDetectorAction extends AbstractAnomalyDetectorAction {
     private static final String VALIDATE_ANOMALY_DETECTOR_ACTION = "validate_anomaly_detector_action";
@@ -72,22 +70,22 @@ public class RestValidateAnomalyDetectorAction extends AbstractAnomalyDetectorAc
     @Override
     public List<ReplacedRoute> replacedRoutes() {
         return ImmutableList
-                .of(
-                        // validate detector
-                        new ReplacedRoute(
-                                RestRequest.Method.POST,
-                                String.format(Locale.ROOT, "%s/%s", AnomalyDetectorPlugin.AD_BASE_DETECTORS_URI, VALIDATE),
-                                RestRequest.Method.POST,
-                                String.format(Locale.ROOT, "%s/%s", AnomalyDetectorPlugin.LEGACY_OPENDISTRO_AD_BASE_URI, VALIDATE)
-                        ),
-                        // validate detector with type
-                        new ReplacedRoute(
-                                RestRequest.Method.POST,
-                                String.format(Locale.ROOT, "%s/%s/{%s}", AnomalyDetectorPlugin.AD_BASE_DETECTORS_URI, VALIDATE, TYPE),
-                                RestRequest.Method.POST,
-                                String.format(Locale.ROOT, "%s/%s/{%s}", AnomalyDetectorPlugin.LEGACY_OPENDISTRO_AD_BASE_URI, VALIDATE, TYPE)
-                        )
-                );
+            .of(
+                // validate detector
+                new ReplacedRoute(
+                    RestRequest.Method.POST,
+                    String.format(Locale.ROOT, "%s/%s", AnomalyDetectorPlugin.AD_BASE_DETECTORS_URI, VALIDATE),
+                    RestRequest.Method.POST,
+                    String.format(Locale.ROOT, "%s/%s", AnomalyDetectorPlugin.LEGACY_OPENDISTRO_AD_BASE_URI, VALIDATE)
+                ),
+                // validate detector with type
+                new ReplacedRoute(
+                    RestRequest.Method.POST,
+                    String.format(Locale.ROOT, "%s/%s/{%s}", AnomalyDetectorPlugin.AD_BASE_DETECTORS_URI, VALIDATE, TYPE),
+                    RestRequest.Method.POST,
+                    String.format(Locale.ROOT, "%s/%s/{%s}", AnomalyDetectorPlugin.LEGACY_OPENDISTRO_AD_BASE_URI, VALIDATE, TYPE)
+                )
+            );
     }
 
     @Override
@@ -102,14 +100,14 @@ public class RestValidateAnomalyDetectorAction extends AbstractAnomalyDetectorAc
         String typesStr = request.param(TYPE);
 
         ValidateAnomalyDetectorRequest validateAnomalyDetectorRequest = new ValidateAnomalyDetectorRequest(
-                detector,
-                typesStr,
-                maxSingleEntityDetectors,
-                maxMultiEntityDetectors,
-                maxAnomalyFeatures,
-                requestTimeout
+            detector,
+            typesStr,
+            maxSingleEntityDetectors,
+            maxMultiEntityDetectors,
+            maxAnomalyFeatures,
+            requestTimeout
         );
         return channel -> client
-                .execute(ValidateAnomalyDetectorAction.INSTANCE, validateAnomalyDetectorRequest, new RestToXContentListener<>(channel));
+            .execute(ValidateAnomalyDetectorAction.INSTANCE, validateAnomalyDetectorRequest, new RestToXContentListener<>(channel));
     }
 }
