+++ src/main/java/org/opensearch/ad/transport/ValidateAnomalyDetectorRequest.java
@@ -30,11 +30,10 @@ import java.io.IOException;
 
 import org.opensearch.action.ActionRequest;
 import org.opensearch.action.ActionRequestValidationException;
+import org.opensearch.ad.model.AnomalyDetector;
 import org.opensearch.common.io.stream.StreamInput;
 import org.opensearch.common.io.stream.StreamOutput;
 import org.opensearch.common.unit.TimeValue;
-
-import org.opensearch.ad.model.AnomalyDetector;
 
 public class ValidateAnomalyDetectorRequest extends ActionRequest {
 
@@ -56,12 +55,12 @@ public class ValidateAnomalyDetectorRequest extends ActionRequest {
     }
 
     public ValidateAnomalyDetectorRequest(
-            AnomalyDetector detector,
-            String typeStr,
-            Integer maxSingleEntityAnomalyDetectors,
-            Integer maxMultiEntityAnomalyDetectors,
-            Integer maxAnomalyFeatures,
-            TimeValue requestTimeout
+        AnomalyDetector detector,
+        String typeStr,
+        Integer maxSingleEntityAnomalyDetectors,
+        Integer maxMultiEntityAnomalyDetectors,
+        Integer maxAnomalyFeatures,
+        TimeValue requestTimeout
     ) {
         this.detector = detector;
         this.typeStr = typeStr;
