+++ src/main/java/org/opensearch/ad/transport/ValidateAnomalyDetectorResponse.java
@@ -29,12 +29,11 @@ package org.opensearch.ad.transport;
 import java.io.IOException;
 
 import org.opensearch.action.ActionResponse;
+import org.opensearch.ad.model.DetectorValidationIssue;
 import org.opensearch.common.io.stream.StreamInput;
 import org.opensearch.common.io.stream.StreamOutput;
 import org.opensearch.common.xcontent.ToXContentObject;
 import org.opensearch.common.xcontent.XContentBuilder;
-
-import org.opensearch.ad.model.DetectorValidationIssue;
 
 public class ValidateAnomalyDetectorResponse extends ActionResponse implements ToXContentObject {
     private DetectorValidationIssue issue;
