+++ src/test/java/org/opensearch/action/admin/indices/mapping/get/IndexAnomalyDetectorActionHandlerTests.java
@@ -71,7 +71,7 @@ import org.opensearch.transport.TransportService;
  * package private
  *
  */
-public class IndexAnomalyDetectorActionHandlerTests extends AnomalyDetectorActionHandlerTestsBase<IndexAnomalyDetectorResponse> {
+public class IndexAnomalyDetectorActionHandlerTests extends AbstractADTest {
     static ThreadPool threadPool;
     private String TEXT_FIELD_TYPE = "text";
     private IndexAnomalyDetectorActionHandler handler;
@@ -94,74 +94,75 @@ public class IndexAnomalyDetectorActionHandlerTests extends AnomalyDetectorActio
     private ADTaskManager adTaskManager;
     private SearchFeatureDao searchFeatureDao;
 
-
     /**
      * Mockito does not allow mock final methods.  Make my own delegates and mock them.
      *
      */
-//    class NodeClientDelegate extends NodeClient {
-//
-//        NodeClientDelegate(Settings settings, ThreadPool threadPool) {
-//            super(settings, threadPool);
-//        }
-//
-//        public <Request extends ActionRequest, Response extends ActionResponse> void execute2(
-//            ActionType<Response> action,
-//            Request request,
-//            ActionListener<Response> listener
-//        ) {
-//            super.execute(action, request, listener);
-//        }
-//
-//    }
-//
-//    @BeforeClass
-//    public static void beforeClass() {
-//        threadPool = new TestThreadPool("IndexAnomalyDetectorJobActionHandlerTests");
-//    }
-//
-//    @AfterClass
-//    public static void afterClass() {
-//        ThreadPool.terminate(threadPool, 30, TimeUnit.SECONDS);
-//        threadPool = null;
-//    }
+    class NodeClientDelegate extends NodeClient {
+
+        NodeClientDelegate(Settings settings, ThreadPool threadPool) {
+            super(settings, threadPool);
+        }
+
+        public <Request extends ActionRequest, Response extends ActionResponse> void execute2(
+            ActionType<Response> action,
+            Request request,
+            ActionListener<Response> listener
+        ) {
+            super.execute(action, request, listener);
+        }
+
+    }
+
+    @BeforeClass
+    public static void beforeClass() {
+        threadPool = new TestThreadPool("IndexAnomalyDetectorJobActionHandlerTests");
+    }
+
+    @AfterClass
+    public static void afterClass() {
+        ThreadPool.terminate(threadPool, 30, TimeUnit.SECONDS);
+        threadPool = null;
+    }
 
     @SuppressWarnings("unchecked")
     @Override
     @Before
     public void setUp() throws Exception {
         super.setUp();
-//
-//        settings = Settings.EMPTY;
-//        clusterService = mock(ClusterService.class);
-//        clientMock = spy(new NodeClient(settings, null));
-//        transportService = mock(TransportService.class);
-//
-//        channel = mock(ActionListener.class);
-//
-//        anomalyDetectionIndices = mock(AnomalyDetectionIndices.class);
-//        when(anomalyDetectionIndices.doesAnomalyDetectorIndexExist()).thenReturn(true);
-//
-//        detectorId = "123";
-//        seqNo = 0L;
-//        primaryTerm = 0L;
-//
-//        WriteRequest.RefreshPolicy refreshPolicy = WriteRequest.RefreshPolicy.IMMEDIATE;
-//
-//        String field = "a";
-//        detector = TestHelpers.randomAnomalyDetectorUsingCategoryFields(detectorId, Arrays.asList(field));
-//
-//        requestTimeout = new TimeValue(1000L);
-//
-//        maxSingleEntityAnomalyDetectors = 1000;
-//
-//        maxMultiEntityAnomalyDetectors = 10;
-//
-//        maxAnomalyFeatures = 5;
-//
-//        method = RestRequest.Method.POST;
-//
-//        adTaskManager = mock(ADTaskManager.class);
+
+        settings = Settings.EMPTY;
+        clusterService = mock(ClusterService.class);
+        clientMock = spy(new NodeClient(settings, null));
+        transportService = mock(TransportService.class);
+
+        channel = mock(ActionListener.class);
+
+        anomalyDetectionIndices = mock(AnomalyDetectionIndices.class);
+        when(anomalyDetectionIndices.doesAnomalyDetectorIndexExist()).thenReturn(true);
+
+        detectorId = "123";
+        seqNo = 0L;
+        primaryTerm = 0L;
+
+        WriteRequest.RefreshPolicy refreshPolicy = WriteRequest.RefreshPolicy.IMMEDIATE;
+
+        String field = "a";
+        detector = TestHelpers.randomAnomalyDetectorUsingCategoryFields(detectorId, Arrays.asList(field));
+
+        requestTimeout = new TimeValue(1000L);
+
+        maxSingleEntityAnomalyDetectors = 1000;
+
+        maxMultiEntityAnomalyDetectors = 10;
+
+        maxAnomalyFeatures = 5;
+
+        method = RestRequest.Method.POST;
+
+        adTaskManager = mock(ADTaskManager.class);
+
+        searchFeatureDao = mock(SearchFeatureDao.class);
 
         handler = new IndexAnomalyDetectorActionHandler(
             clusterService,
