+++ src/test/java/org/opensearch/ad/TestHelpers.java
@@ -73,17 +73,17 @@ import org.opensearch.ad.model.ADTaskType;
 import org.opensearch.ad.model.AnomalyDetector;
 import org.opensearch.ad.model.AnomalyDetectorExecutionInput;
 import org.opensearch.ad.model.AnomalyDetectorJob;
-import org.opensearch.ad.model.ValidationAspect;
-import org.opensearch.ad.model.DetectorValidationIssueType;
-import org.opensearch.ad.model.DetectorValidationIssue;
 import org.opensearch.ad.model.AnomalyResult;
 import org.opensearch.ad.model.DetectionDateRange;
 import org.opensearch.ad.model.DetectorInternalState;
+import org.opensearch.ad.model.DetectorValidationIssue;
+import org.opensearch.ad.model.DetectorValidationIssueType;
 import org.opensearch.ad.model.Entity;
 import org.opensearch.ad.model.Feature;
 import org.opensearch.ad.model.FeatureData;
 import org.opensearch.ad.model.IntervalTimeConfiguration;
 import org.opensearch.ad.model.TimeConfiguration;
+import org.opensearch.ad.model.ValidationAspect;
 import org.opensearch.ad.settings.AnomalyDetectorSettings;
 import org.opensearch.client.AdminClient;
 import org.opensearch.client.Client;
@@ -311,8 +311,6 @@ public class TestHelpers {
             user
         );
     }
-
-
 
     public static AnomalyDetector randomDetector(List<Feature> features, String indexName, int detectionIntervalInMinutes, String timeField)
         throws IOException {
@@ -1248,11 +1246,11 @@ public class TestHelpers {
 
     public static DetectorValidationIssue randomDetectorValidationIssue() {
         DetectorValidationIssue issue = new DetectorValidationIssue(
-                ValidationAspect.DETECTOR,
-                DetectorValidationIssueType.NAME,
-                randomAlphaOfLength(5),
-                null,
-                null
+            ValidationAspect.DETECTOR,
+            DetectorValidationIssueType.NAME,
+            randomAlphaOfLength(5),
+            null,
+            null
         );
         return issue;
     }
